(Added) new 'main' and VTS
Added for main: Boundary conditions in arbitrary for now is available, any linear combination of function and it's derivatives. Also added numerical schemes for derivatives up to 4 order.

Added for VTS: such main solver added to VTS.

Important problem: we should fix the number of dots on setka with respect to degree of initial diff_eq and the higher derivative in boundary at the right dot